Force feats casting inside flex_gemm wrapper

+++ trellis2/modules/sparse/conv/conv_flex_gemm.py
@@ -44,7 +44,7 @@ def sparse_conv3d_forward(self, x: SparseTensor) -> SparseTensor:
     neighbor_cache = x.get_spatial_cache(neighbor_cache_key)
     
     out, neighbor_cache_ = sparse_submanifold_conv3d(
-        x.feats,
+        x.feats.to(self.weight.dtype),
         x.coords,
         torch.Size([*x.shape, *x.spatial_shape]),
         self.weight,
